Refactoring to make duplication less necessary
