+++ instaseis/server/instaseis_request.py
@@ -15,6 +15,7 @@ from abc import ABCMeta, abstractmethod
 import obspy
 import tornado
 from ..base_instaseis_db import _get_seismogram_times
+from .. import Receiver
 
 from .. import __version__
 
@@ -312,13 +313,9 @@ class InstaseisTimeSeriesHandler(with_metaclass(ABCMeta,
 
         # Any single...
         if hasattr(source, "latitude"):
-            src_latitude = source.latitude
-            src_longitude = source.longitude
             src_depth_in_m = source.depth_in_m
         # Or a finite source...
         else:
-            src_latitude = source.hypocenter_latitude
-            src_longitude = source.hypocenter_longitude
             src_depth_in_m = source.hypocenter_depth_in_m
 
         # XXX: Will have to be changed once we have a database recorded for
@@ -385,3 +382,92 @@ class InstaseisTimeSeriesHandler(with_metaclass(ABCMeta,
             return
 
         return starttime, endtime
+
+    def get_receivers(self, args):
+        # Already checked before - just make sure the settings are valid.
+        assert (args.receiverlatitude is not None and
+                args.receiverlongitude is not None) or \
+               (args.network and args.station)
+
+        receivers = []
+
+        # Construct either a single receiver object.
+        if args.receiverlatitude is not None:
+            try:
+                receiver = Receiver(latitude=args.receiverlatitude,
+                                    longitude=args.receiverlongitude,
+                                    network=args.networkcode,
+                                    station=args.stationcode,
+                                    depth_in_m=args.receiverdepthinmeters)
+            except:
+                msg = ("Could not construct receiver with passed parameters. "
+                       "Check parameters for sanity.")
+                raise tornado.web.HTTPError(400, log_message=msg, reason=msg)
+            receivers.append(receiver)
+        # Or a list of receivers.
+        elif args.network is not None and args.station is not None:
+            networks = args.network.split(",")
+            stations = args.station.split(",")
+
+            coordinates = self.application.station_coordinates_callback(
+                networks=networks, stations=stations)
+
+            if not coordinates:
+                msg = "No coordinates found satisfying the query."
+                raise tornado.web.HTTPError(
+                    404, log_message=msg, reason=msg)
+
+            for station in coordinates:
+                try:
+                    receivers.append(Receiver(
+                        latitude=station["latitude"],
+                        longitude=station["longitude"],
+                        network=station["network"],
+                        station=station["station"],
+                        depth_in_m=0))
+                except:
+                    msg = ("Could not construct receiver with passed "
+                           "parameters. Check parameters for sanity.")
+                    raise tornado.web.HTTPError(400, log_message=msg,
+                                                reason=msg)
+        return receivers
+
+    def validate_receiver_parameters(self, args):
+        """
+        Useful for routes that use single receivers.
+        """
+        # The networkcode and stationcode parameters have a maximum number
+        # of letters.
+        if args.stationcode and len(args.stationcode) > 5:
+            msg = "'stationcode' must have 5 or fewer letters."
+            raise tornado.web.HTTPError(400, log_message=msg, reason=msg)
+
+        if args.networkcode and len(args.networkcode) > 2:
+            msg = "'networkcode' must have 2 or fewer letters."
+            raise tornado.web.HTTPError(400, log_message=msg, reason=msg)
+
+        # Figure out who the station coordinates are specified.
+        direct_receiver_settings = [
+            i is not None
+            for i in (args.receiverlatitude, args.receiverlongitude)]
+        query_receivers = [i is not None for i in (args.network, args.station)]
+        if any(direct_receiver_settings) and any(query_receivers):
+            msg = ("Receiver coordinates can either be specified by passing "
+                   "the coordinates, or by specifying query parameters, "
+                   "but not both.")
+            raise tornado.web.HTTPError(400, log_message=msg, reason=msg)
+        elif not(all(direct_receiver_settings) or all(query_receivers)):
+            msg = ("Must specify a full set of coordinates or a full set of "
+                   "receiver parameters.")
+            raise tornado.web.HTTPError(400, log_message=msg, reason=msg)
+
+        # Should not happen.
+        assert not (all(direct_receiver_settings) and all(query_receivers))
+
+        # Make sure that the station coordinates callback is available if
+        # needed. Otherwise raise a 404.
+        if all(query_receivers) and \
+                not self.application.station_coordinates_callback:
+            msg = ("Server does not support station coordinates and thus no "
+                   "station queries.")
+            raise tornado.web.HTTPError(404, log_message=msg, reason=msg)
